Refactor: Board CRUD 리팩토링 #4

+++ src/board/board.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { CreateBoardDto } from './dto/create-board.dto';
 import { UpdateBoardDto } from './dto/update-board.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -22,12 +22,20 @@ export class BoardService {
     private readonly emailService: EmailService,
   ) {}
 
-  async create(createBoardDto: CreateBoardDto) {
-    const { title, backgroundColor, adminId } = createBoardDto;
+  async create(userId: number, createBoardDto: CreateBoardDto) {
+    const { title, backgroundColor, description } = createBoardDto;
+    // adminId를 userId로 등록
     const board = await this.boardRepository.save({
-      adminId,
+      adminId: userId,
       title,
       backgroundColor,
+      description,
+    });
+
+    const boardMember = await this.boardMemberRepository.save({
+      boardId: board.id,
+      memberId: userId,
+      memberType: BoardMemberType.ADMIN,
     });
     return board;
   }
@@ -43,10 +51,25 @@ export class BoardService {
         id,
       },
     });
+    if (!board) {
+      throw new NotFoundError('보드가 존재하지 않습니다.');
+    }
     return board;
   }
 
-  async update(id: number, updateBoardDto: UpdateBoardDto) {
+  async update(id: number, userId: number, updateBoardDto: UpdateBoardDto) {
+    const board = await this.boardRepository.findOne({
+      where: {
+        id,
+      },
+    });
+    if (!board) {
+      throw new NotFoundError('보드가 존재하지 않습니다.');
+    }
+    // board의 admin만 수정 가능
+    if (userId !== board.adminId) {
+      throw new UnauthorizedException('수정 권한이 없습니다.');
+    }
     const updatingBoard = await this.boardRepository.update(id, updateBoardDto);
     const updatedBoard = await this.boardRepository.findOne({
       where: {
@@ -56,8 +79,20 @@ export class BoardService {
     return updatedBoard;
   }
 
-  async delete(id: number) {
-    const deletingBoard = await this.boardRepository.delete(id)
+  async delete(id: number, userId: number) {
+    const board = await this.boardRepository.findOne({
+      where: {
+        id,
+      },
+    });
+    if (!board) {
+      throw new NotFoundError('보드가 존재하지 않습니다.');
+    }
+    // board의 admin만 삭제 가능
+    if (userId !== board.adminId) {
+      throw new UnauthorizedException('삭제 권한이 없습니다.');
+    }
+    const deletingBoard = await this.boardRepository.delete(id);
     return;
   }
